Move location data fetching into a helper function.
