+++ src/components/calculator/PrevalenceControls.tsx
@@ -9,8 +9,9 @@ import ControlLabel from './controls/ControlLabel'
 import { LocationPrevalenceDetails } from './prevalence/LocationPrevalenceDetails'
 import { ManualPrevalenceDetails } from './prevalence/ManualPrevalenceDetails'
 import { CalculatorData } from 'data/calculate'
-import { Locations, PrevalenceDataDate } from 'data/location'
+import { Locations } from 'data/location'
 import 'components/calculator/styles/PrevalenceControls.scss'
+import { dataForLocation } from 'data/locationHelpers'
 
 interface Option {
   label: string
@@ -248,48 +249,4 @@ export const PrevalenceControls: React.FunctionComponent<{
       )}
     </React.Fragment>
   )
-}
-
-interface PrevalanceData {
-  population: string
-  casesPastWeek: number
-  casesIncreasingPercentage: number
-  positiveCasePercentage: number | null
-  prevalanceDataDate: Date
-  completeVaccinations: number | null
-  unvaccinatedPrevalenceRatio: number | null
-  averageFullyVaccinatedMultiplier: number | null
-}
-
-function dataForLocation(location: string): PrevalanceData {
-  const locationData = Locations[location]
-
-  if (locationData) {
-    return {
-      population: locationData.population,
-      casesPastWeek: locationData.casesPastWeek,
-      casesIncreasingPercentage:
-        Math.round(locationData.casesIncreasingPercentage * 10) / 10,
-      positiveCasePercentage:
-        locationData.positiveCasePercentage === null
-          ? null
-          : Math.round(locationData.positiveCasePercentage * 10) / 10,
-      prevalanceDataDate: new Date(PrevalenceDataDate),
-      completeVaccinations: locationData.completeVaccinations,
-      unvaccinatedPrevalenceRatio: locationData.unvaccinatedPrevalenceRatio,
-      averageFullyVaccinatedMultiplier:
-        locationData.averageFullyVaccinatedMultiplier,
-    }
-  }
-
-  return {
-    population: '',
-    casesPastWeek: 0,
-    casesIncreasingPercentage: 0,
-    positiveCasePercentage: 0,
-    prevalanceDataDate: new Date(),
-    completeVaccinations: null,
-    unvaccinatedPrevalenceRatio: null,
-    averageFullyVaccinatedMultiplier: null,
-  }
 }
